+++ es_29_pag_73.py
@@ -5,16 +5,16 @@ while True:
     if nome=="0":
         break  
     citta.append(nome)  
-    temp_min=int(input("Quanto è stata la temperatura minima oggi [°C]? "))
+    temp_min=float(input("Quanto è stata la temperatura minima oggi [°C]? "))
     citta.append(temp_min)  
-    temp_max=int(input("Quanto è stata la temperatura massima oggi [°C]? "))
+    temp_max=float(input("Quanto è stata la temperatura massima oggi [°C]? "))
     citta.append(temp_max)  
     lista_citta.append(citta)
 
 citta_cont=0
 lista_citta_escursione_superiore=[]
 print()
-escursione_prefiss=int(input("Qual è l'escursione termica da verificare? "))
+escursione_prefiss=float(input("Qual è l'escursione termica da verificare? "))
 for c in lista_citta:
     if (c[2]-c[1])>escursione_prefiss:
         citta_cont+=1
